fuel_monitoring/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.1
tzdata==2024.2
djangorestframework_simplejwt==4.7.2
